CODE/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
altair==4.1.0             # via -r requirements.in, streamlit
argon2-cffi==20.1.0       # via notebook
astor==0.8.1              # via streamlit
async-generator==1.10     # via nbclient
attrs==20.3.0             # via fiona, jsonschema, pytest
backcall==0.2.0           # via ipython
base58==2.0.1             # via streamlit
bleach==3.2.1             # via nbconvert
blinker==1.4              # via streamlit
blis==0.4.1               # via spacy, thinc
booste==0.2.3             # via -r requirements.in
boto3==1.16.13            # via streamlit
botocore==1.19.13         # via boto3, s3transfer, streamlit
cachetools==4.1.1         # via streamlit
catalogue==1.0.0          # via spacy, thinc
certifi==2020.6.20        # via booste, pyproj, requests
cffi==1.14.3              # via argon2-cffi
chardet==3.0.4            # via booste, requests
click-plugins==1.1.1      # via fiona
click==7.1.2              # via click-plugins, cligj, fiona, nltk, streamlit
cligj==0.7.0              # via fiona
coverage==5.3             # via pytest-cov
cymem==2.0.4              # via preshed, spacy, thinc
decorator==4.4.2          # via ipython, validators
defusedxml==0.6.0         # via nbconvert
entrypoints==0.3          # via altair, nbconvert
enum-compat==0.0.3        # via streamlit
fiona==1.8.17             # via geopandas
funcy==1.15               # via pyldavis
future==0.18.2            # via pyldavis
gensim==3.8.3             # via -r requirements.in
geopandas==0.8.1          # via -r requirements.in
gitdb==4.0.5              # via gitpython
gitpython==3.1.11         # via streamlit
graphviz==0.14.2          # via -r requirements.in
idna==2.10                # via booste, requests
iniconfig==1.1.1          # via pytest
ipykernel==5.3.4          # via ipywidgets, jupyter, jupyter-console, notebook, pydeck, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==7.19.0           # via ipykernel, ipywidgets, jupyter-console
ipywidgets==7.5.1         # via jupyter, pydeck
isodate==0.6.0            # via rdflib
jedi==0.17.2              # via ipython
jinja2==2.11.2            # via altair, nbconvert, notebook, pydeck, pyldavis
jmespath==0.10.0          # via boto3, botocore
joblib==0.17.0            # via -r requirements.in, nltk, pyldavis, scikit-learn
jsonschema==3.2.0         # via altair, nbformat
jupyter-client==6.1.7     # via ipykernel, jupyter-console, nbclient, notebook, qtconsole
jupyter-console==6.2.0    # via jupyter
jupyter-core==4.6.3       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0            # via -r requirements.in
jupyterlab-pygments==0.1.2  # via nbconvert
markupsafe==1.1.1         # via jinja2
mistune==0.8.4            # via nbconvert
munch==2.5.0              # via fiona
murmurhash==1.0.4         # via preshed, spacy, thinc
nbclient==0.5.1           # via nbconvert
nbconvert==6.0.7          # via jupyter, notebook
nbformat==5.0.8           # via ipywidgets, nbclient, nbconvert, notebook
nest-asyncio==1.4.2       # via nbclient
nltk==3.5                 # via -r requirements.in
notebook==6.1.5           # via jupyter, widgetsnbextension
numexpr==2.7.1            # via pyldavis
numpy==1.19.4             # via -r requirements.in, altair, blis, booste, gensim, numexpr, pandas, pyarrow, pydeck, pyldavis, scikit-learn, scipy, spacy, streamlit, thinc
packaging==20.4           # via bleach, pytest, streamlit
pandas==1.1.4             # via -r requirements.in, altair, geopandas, pyldavis, streamlit, vega-datasets
pandocfilters==1.4.3      # via nbconvert
parso==0.7.1              # via jedi
pathtools==0.1.2          # via watchdog
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==8.0.1             # via streamlit
plac==1.1.3               # via spacy, thinc
plotly==4.12.0            # via -r requirements.in
pluggy==0.13.1            # via pytest
preshed==3.0.4            # via spacy, thinc
prometheus-client==0.8.0  # via notebook
prompt-toolkit==3.0.8     # via ipython, jupyter-console
protobuf==3.13.0          # via streamlit
ptyprocess==0.6.0         # via pexpect, terminado
py==1.9.0                 # via pytest
pyarrow==2.0.0            # via streamlit
pycparser==2.20           # via cffi
pydeck==0.5.0             # via streamlit
pygments==2.7.2           # via ipython, jupyter-console, jupyterlab-pygments, nbconvert, qtconsole
pyldavis==2.1.2           # via -r requirements.in
pyparsing==2.4.7          # via packaging, rdflib
pyproj==3.0.0.post1       # via geopandas
pyrsistent==0.17.3        # via jsonschema
pytest-cov==2.10.1        # via -r requirements.in
pytest-mock==3.3.1        # via -r requirements.in
pytest==6.1.2             # via -r requirements.in, pyldavis, pytest-cov, pytest-mock
python-dateutil==2.8.1    # via botocore, jupyter-client, pandas, streamlit
python-dotenv==0.15.0     # via -r requirements.in
pytz==2020.4              # via pandas, tzlocal
pyzmq==19.0.2             # via jupyter-client, notebook, qtconsole
qtconsole==4.7.7          # via jupyter
qtpy==1.9.0               # via qtconsole
rdflib==5.0.0             # via -r requirements.in
regex==2020.11.13         # via nltk
requests==2.24.0          # via booste, smart-open, spacy, streamlit
retrying==1.3.3           # via plotly
s3transfer==0.3.3         # via boto3
scikit-learn==0.23.2      # via sklearn
scipy==1.5.4              # via gensim, pyldavis, scikit-learn
send2trash==1.5.0         # via notebook
shapely==1.7.1            # via geopandas
six==1.15.0               # via argon2-cffi, bleach, fiona, gensim, isodate, jsonschema, munch, packaging, plotly, protobuf, python-dateutil, rdflib, retrying, validators
sklearn==0.0              # via -r requirements.in
smart-open==3.0.0         # via gensim
smmap==3.0.4              # via gitdb
spacy==2.3.2              # via -r requirements.in
srsly==1.0.4              # via spacy, thinc
streamlit==0.70.0         # via -r requirements.in
terminado==0.9.1          # via notebook
testpath==0.4.4           # via nbconvert
thinc==7.4.1              # via spacy
threadpoolctl==2.1.0      # via scikit-learn
toml==0.10.2              # via pytest, streamlit
toolz==0.11.1             # via altair
tornado==6.1              # via ipykernel, jupyter-client, notebook, streamlit, terminado
tqdm==4.52.0              # via nltk, spacy, thinc
traitlets==5.0.5          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbclient, nbconvert, nbformat, notebook, pydeck, qtconsole
tzlocal==2.1              # via streamlit
urllib3==1.25.11          # via booste, botocore, requests
validators==0.18.1        # via streamlit
vega-datasets==0.8.0      # via -r requirements.in
wasabi==0.8.0             # via spacy, thinc
watchdog==0.10.3          # via streamlit
wcwidth==0.2.5            # via prompt-toolkit
webencodings==0.5.1       # via bleach
wheel==0.35.1             # via pyldavis
widgetsnbextension==3.5.1  # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# setuptools
